case 3/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
python-dotenv
pydantic